validate get_payload signature inline, remove sub worker
