+++ crates/relay/src/lib.rs
@@ -28,7 +28,7 @@ pub use crate::{
     auctioneer::{
         Auctioneer, AuctioneerHandle, BidSorter, BlockSimRequest, Context, Event,
         InternalBidSubmission, PayloadEntry, RegWorker, RegWorkerHandle, SimulatorClient,
-        SimulatorManager, SimulatorRequest, SlotData, SubWorker, SubmissionPayload,
+        SimulatorManager, SimulatorRequest, SlotData, SubmissionPayload,
     },
     beacon::start_beacon_client,
     bid_decoder::{DecoderTile, SubmissionDataWithSpan},
